feat: Add backgroundColor function
This function sets the background color. This is used less than one
might think, because generally we want to leave the background untouched
so that it doesn't interfere with terminal transparency, and for most
circumstances we just use color inversion instead.

At time of writing, this is only really used for the grid-display
feature.